refactor, fix size issue
